requirements.txt
streamlit
pandas
fpdf2
openpyxl
Pillow